NBZIMM-package {NBZIMM}	R Documentation

Negative Binomial and Zero-Inflated Mixed Models

Description

This package provides functions for setting up and fitting negative binomial mixed models and zero-inflated negative binomial and Gaussian mixed models. These functions allow for multiple and correlated group-specific (random) effects and various types of within-group correlation structures as described in the core package nlme, and return objects that can be summarized by functions in nlme. The methods can be used to analyze overdispersed and zero-inflated count or continuous responses with multilevel data structures (for example, clustered data and longitudinal studies).

Details

Package: 	 NBZIMM

Type: 	 Package

Version: 	 1.0

Date: 	 2019-02-01

License: 	 GPL

LazyLoad: 	 yes

Author(s)

Nengjun Yi, nyi@uab.edu

[Package NBZIMM version 1.0 Index]