Romero {NBZIMM}	R Documentation

Longitudinal Microbiome Count Data of Pregnant and Non-pregnant Women (Romero et al. 2014)

Description

Romero et al. (2014) published a longitudinal study of the vaginal microbiota between normal pregnant women and non-pregnant women.

Usage


data(Romero)

Details

This case-control longitudinal study collected 22 pregnant women who delivered at term and 32 non-pregnant women, each measured at multiple time points. The data consist of two data components: OTU and SampleData. OTU contains microbiome count data for 900 samples with 143 taxa; SampleData contains data of sample variables for all the samples with 9 variables, including subject ID, pregnant status, total sequencing read, age, race, etc.

References

Romero, R. et al. (2014) The composition and stability of the vaginal microbiota of normal pregnant women is different from that of non-pregnant women. Microbiome 2:4.

[Package NBZIMM version 1.0 Index]