fixed {NBZIMM}	R Documentation

Extracting and Summarizing Fixed Effects

Description

This function is to extract the estimates, standard deviations and p-values of fixed effects from the output of glmm.nb, lme, glmm.zinb, lme.zig, or mms.

Usage


fixed(object) 

fixed.nb(object)

fixed.lme(object)

fixed.zi(object)

fixed.mms(object)

Arguments

object
an object from glmm.nb, lme, glmm.zinb, lme.zig, or mms.

Value

This function returns the following values for the distribution part (dist) and/or the zero-inflation part (zero):

Estimate
intercept and coefficient estimates;

Std. Error
standard errors of estimates;

pvalue
p values for testing coefficients;

Author(s)

Nengjun Yi, nyi@uab.edu

Examples


see examples in glmm.nb, glmm.zinb, lme.zig, and mms 

[Package NBZIMM version 1.0 Index]