get.fixed {NBZIMM}	R Documentation

Extract Fixed Effects for A Response or Variable from a 'mms' Output

Description

This function extracts the estimates, standard deviations and p-values of fixed effects for a response or variable from the output of mms.

Usage

  
get.fixed(object, part = c("dist", "zero"), vr.name, sort.p = FALSE)

Arguments

object
an object from mms.

part
distribution part or zero-inflation part.

vr.name
name of a variable or response.

sort.p
sort by the p-values into ascending order.

Value

A matrix consists of the estimates, standard deviations and p-values of fixed effects for the response or variable.

Author(s)

Nengjun Yi, nyi@uab.edu

See Also

mms

Examples


see examples in mms 


[Package NBZIMM version 1.0 Index]