glmm.nb {NBZIMM}	R Documentation

Negative Binomial Mixed Models

Description

This function sets up and fits negative binomial mixed models for analyzing overdispersed count responses with multilevel data structures (for example, clustered data and longitudinal studies).

Usage

  
glmm.nb(fixed, random, data, correlation,  
        niter = 30, epsilon = 1e-05, verbose = TRUE, ...)  

Arguments

fixed, random, data, correlation
These arguments are the same as in the function lme in the package nlme.

niter
maximum number of iterations.

epsilon
positive convergence tolerance.

verbose
logical. If TRUE, print out number of iterations and computational time.

...
further arguments for lme.

Details

This function is an alteration of the function glmmPQL in the package MASS, which fits generalized linear mixed models using Penalized Quasi-Likelihood and works by repeated calls to the function lme in the package nlme. It sets up and fits negative binomial mixed-effects models, which are the standard models for analyzing overdispersed count responses in mutilevel study designs, for example, clustered and longitudinal studies. The function allows for multiple and correlated group-specific (random) effects (the argument random) and various types of within-group correlation structures (the argument correlation) described by corStruct in the package nlme.

Value

A fitted model object of class lme inheriting from lme, which can be summarized by functions in the package nlme. The object also contains the estimate of the dispersion parameter theta.

Author(s)

Nengjun Yi, nyi@uab.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

Venables, W. N. and Ripley, B. D. (2002) "Modern Applied Statistics with S". Fourth edition. Springer.

Xinyan Zhang, Himel Mallick, Xiangqin Cui, Andrew K. Benson, and Nengjun Yi (2017) Negative Binomial Mixed Models for Analyzing Microbiome Count Data. BMC Bioinformatics 18(1):4.

Xinyan Zhang, Yu-Fang Pei, Lei Zhang, Boyi Guo, Amanda Pendegraft, Wenzhuo Zhuang and Nengjun Yi (2018) Negative Binomial Mixed Models for Analyzing Longitudinal Microbiome Data. Frontiers in Microbiology.

See Also

lme, glmmPQL

Examples



library(NBZIMM)

data(Romero)
names(Romero)

otu = Romero$OTU; dim(otu)
sam = Romero$SampleData; dim(sam)
colnames(sam)

N = sam[, "Total.Read.Counts"]        
Days = sam$GA_Days; Days = scale(Days)
Age = sam$Age; Age = scale(Age)
Race = sam$Race
preg = sam$pregnant; table(preg)

subject = sam[, "Subect_ID"]; table(subject)

non = nonzero(y = otu, total = N, plot = F)
nonzero.p = non[[1]]

y = otu[, names(nonzero.p)[1]]

f = glmm.nb(y ~ Days + Age + Race + preg + offset(log(N)), random = ~ 1 | subject)
summary(f)
fixed(f)


library(lme4)  # compared with lme4
f0 = glmer.nb(y ~ Days + Age + Race + preg + offset(log(N)) + (1 | subject))   
summary(f0)


[Package NBZIMM version 1.0 Index]