glmm.zinb {NBZIMM}	R Documentation

Zero-Inflated Negative Binomial Mixed Models

Description

This function sets up and fits zero-inflated negative binomial mixed models for analyzing zero-inflated count responses with multilevel data structures (for example, clustered data and longitudinal studies).

Usage

  
glmm.zinb(fixed, random, data, correlation, zi.random = FALSE, 
        niter = 30, epsilon = 1e-05, verbose = TRUE, ...)  

Arguments

fixed
symbolic description of the fixed-effects part of the model, see details.

random, data, correlation
These arguments are the same as in the function lme in the package nlme.

zi.random
logical. If TRUE, include the random effect specified by random in the zero-inflation part.

niter
maximum number of iterations.

epsilon
positive convergence tolerance.

verbose
logical. If TRUE, print out number of iterations and computational time.

...
further arguments for lme.

Details

Zero-inflated models are two-component mixture models combining a point mass at zero with a negative binomial distribution for count response. Thus, there are two sources of zeros: zeros may come from either the point mass at zero or the negative binomial distribution. For modeling the unobserved state, a logistic model (logistic mixed model if zi.random = TRUE) is used.

The argument fixed is to specify the fixed-effects part in both components of the model: fixed = y ~ x | z giving the fixed-effects part of the distribution model y ~ x conditional on (|) the zero-inflation model ~ z. A different (or same) set of fixed-effects covariates could be used for the distribution component and zero-inflation component. If fixed = y ~ x, then the same covariates are employed in both components, equivalent to y ~ x | x. The simplest zero-inflation model only includes an intercept: y ~ x | 1, and thus all zeros have the same probability of belonging to the zero component. Offsets can be specified in both components of the model : y ~ x + offset(x1) | z + offset(z1).

The function is an alteration of the function glmmPQL in the package MASS, which fits generalized linear mixed models using Penalized Quasi-Likelihood and works by repeated calls to the function lme in the package nlme. It incorporates EM algorithms into the procedure of glmmPQL to fit zero-inflated mixed models for analyzing zero-inflated count responses in mutilevel study designs, for example, clustered and longitudinal studies. The function allows for multiple and correlated group-specific (random) effects (the argument random) and various types of within-group correlation structures (the argument correlation) described by corStruct in the package nlme.

Missing data may not be properly handled in some situations and thus should be removed before the analysis.

Value

A fitted model object of class lme inheriting from lme, which can be summarized by functions in the package nlme for the count distribution part.

The object contains additional components for the zero-inflation part:

theta
the estimate of the dispersion parameter;

zero.prob
the zero-state probabilities;

zero.indicator
the conditional expectations of the zero indicators;

xz
the design matrix of the zero-inflation part;

offsetz
the offset of the zero-inflation part;

fit.zero
the fitted logistic (mixed) model for the zero-inflation part;

Author(s)

Nengjun Yi, nyi@uab.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

Venables, W. N. and Ripley, B. D. (2002) "Modern Applied Statistics with S". Fourth edition. Springer.

Xinyan Zhang, Himel Mallick, Xiangqin Cui, Andrew K. Benson, and Nengjun Yi (2017) Negative Binomial Mixed Models for Analyzing Microbiome Count Data. BMC Bioinformatics 18(1):4.

Xinyan Zhang, Yu-Fang Pei, Lei Zhang, Boyi Guo, Amanda Pendegraft, Wenzhuo Zhuang and Nengjun Yi (2018) Negative Binomial Mixed Models for Analyzing Longitudinal Microbiome Data. Frontiers in Microbiology.

See Also

lme, glmmPQL, glmm.nb, lme.zig

Examples



library(NBZIMM)

data(Romero)
names(Romero)

otu = Romero$OTU; dim(otu)
sam = Romero$SampleData; dim(sam)
colnames(sam)

N = sam[, "Total.Read.Counts"]        
Days = sam$GA_Days; Days = scale(Days)
Age = sam$Age; Age = scale(Age)
Race = sam$Race
preg = sam$pregnant; table(preg)

subject = sam[, "Subect_ID"]; table(subject)

non = nonzero(y = otu, total = N, plot = F)
nonzero.p = non[[1]]

y = otu[, names(nonzero.p)[1]]

f = glmm.zinb(y ~ Days + Age + Race + preg + offset(log(N)) | 1, random = ~ 1 | subject) 
summary(f)
fixed(f)
summary(f$fit.zero)


[Package NBZIMM version 1.0 Index]