heat.p {NBZIMM}	R Documentation

ggplot2 heatmap of p values

Description

Visualizing p values and positive coefficients for many responses and variables.

Usage


heat.p(df, p.breaks = c(0.001, 0.01, 0.05), 
      colors = c("darkblue", "blue", "lightblue", "grey"),
      title = NULL, subtitle = NULL,
      filter = c(FALSE, FALSE), x.size, y.size, symbol.size,
      zigzag = c(FALSE, FALSE), abbrv = c(FALSE, FALSE), 
      margin = c(1, 1), legend = TRUE) 

Arguments

df
data frame, giving response names, variable names, coefficient estimates and p values for many responses and variables. The column names must include "responses", "variables", "Estimate" and "pvalue".

p.breaks
a vector (not including 0 and 1), partitioning p values into intervals.

colors
different colors for different p value intervals.

title
The text for title for the plot.

subtitle
The text for subtitle below the title.

filter
TRUE or FALSE. consist of two elements, the first is for responses (y-axis) and the second is for variables (x-axis). If TRUE, remove responses or variables with all p values larger than the maximun value of p.breaks.

x.size, y.size
size for names of x-axis or y-axis.

symbol.size
size for the symbol "+" for positive estimates (only those with p values smaller than the maximun value of p.breaks).

zigzag
TRUE or FALSE. consist of two elements, the first is for responses (y-axis) and the second is for variables (x-axis). If TRUE, re-arrange y-axis or x-axis labels in a zigzag shape.

abbrv
TRUE or FALSE. consist of two elements, the first is for responses (y-axis) and the second is for variables (x-axis). If TRUE, abbreviate y-axis or x-axis labels.

marigin
a numeric vector of two values for the margins in centimeter. the first is for the left and the second is for the bottom.

legend
TRUE or FALSE. show the legend for p values or not.

Value

ggplot2 object

Author(s)

Boyi Guo, boyiguo1@uab.edu; Xinyan Zhang, abbyyan3@gmail.com; Nengjun Yi, nyi@uab.edu

Examples


see examples in mms 

[Package NBZIMM version 1.0 Index]