nonzero {NBZIMM}	R Documentation

Proportions of Non-Zero Values

Description

This simple function calculates the proportion of non-zero values for each response and plots the proportions.

Usage


nonzero(y, total, plot = FALSE)

Arguments

y
a matrix of responses; each column is a response and rows are samples.

total
optional. It is total number (total reads in microbiome data). If not provided, the number is calculated as the column sum of y.

plot
plot zero proportion and total number or not.

Value

ordered proportions of non-zero values for responses: nonzero.p, total numbers for all samples: total, mean and standard deviation of total numbers: total_mean_sd, and plots of ordered proportions of zero values and total numbers.

Author(s)

Nengjun Yi, nyi@uab.edu

Examples



library(NBZIMM)

data(Romero)
names(Romero)
otu = Romero$OTU; dim(otu)
sam = Romero$SampleData; dim(sam)
colnames(sam)
N = sam[, "Total.Read.Counts"]  # total reads

non = nonzero(y = otu, total = N, plot = T)
non


[Package NBZIMM version 1.0 Index]