plot.fixed {NBZIMM}	R Documentation

Plotting Estimates, Standard Deviations and P-values of Fixed Effects

Description

Plotting estimates, standard deviations and p-values of fixed effects.

Usage


plot.fixed(res, threshold = 0.05, main = " ", col.pts = "black", gap = 0, 
          show.all.vars = FALSE, show.pvalues = TRUE, 
          cex.main = 0.9, xlim = NULL, cex.axis = 0.8, cex.var = 0.8, cex.pts = 1, 
          pch.pts = 20, type = "p", lwd = 1, lty = 1, line = 0,  
          add = FALSE) 

Arguments

res
A matrix of the estimates, standard deviations and p-values of fixed effects for the response or variable.

threshold
a p-value to determine variables to be shown: variables with p-value < threshold.

col.pts
color of points and segments, default is black. It can be a vector with two elements, which will use specified colors for different variables determined by the augument threshold.

gap
a value for the distance between two flanking significant variables.

main, cex.main, xlim, type, lwd, lty, line
These arguments are the same as in plot and par.

show.all.vars
logical. If TRUE,show names of all variables. The default is FALSE.

show.pvalues
logical. If TRUE, show p-values for variables. The default is TRUE. The p-values are shown on the right side of the plot.

cex.var
the fontsize of the varible names, default = 0.8.

cex.pts
the size of points, default = 1.

pch.pts
symbol of points, default is solid dot.

add
logical. if TRUE, plot over the existing plot. The default is FALSE.

Details

This function plots the estimates of coefficients, intervals and p-values for a response or variable. It uses different colors to distinguish between significant and insignificant variables based on a threshold.

Author(s)

Nengjun Yi, nyi@uab.edu

See Also

plot, par

Examples


see examples in mms 

[Package NBZIMM version 1.0 Index]